feat: Created a new branch for parsing raw JSON data
Found out that there's a ResponseBody class that's not within
Retrofit, but it's actually within OkHttp (which Retrofit is built
off of). Using this class, you can get the raw HTTP response from
network calls. I created a new branch called RawJSONData which
includes a separate solution that parses the raw JSON data,
instead of adding some of the additional overhead that was
involved to make Retrofit work.